Refactor to allow app to be started without babel.
